Add files via upload
Jake's plotter - altered value of dr a couple comments truncated for professionalism.
Additional file D_function_intiial_conc.py. This will be used to return a constant value of the diffusion constant D, dependent upon the initial concentration. This is currently a placeholder file that returns 1.0, so it can be implemented prior to final development if desired.